change in types

+++ src/emulator/branch.c
@@ -62,7 +62,7 @@ void branch(ARM* arm, int instruction) {
 
     switch (type) {
         case UNCONDITIONAL: {
-            int32_t simm26 = getBitsAt(instruction, BR_SIMM26_START, SIMM26_LEN);
+            int64_t simm26 = getBitsAt(instruction, BR_SIMM26_START, SIMM26_LEN);
             int64_t offset = simm26 * BYTES_IN_WORD;
             // Branch to address encoded by literal
             arm->pc += offset;
@@ -78,7 +78,7 @@ void branch(ARM* arm, int instruction) {
             break;
         }
         case CONDITIONAL: {
-            int32_t simm19 = getBitsAt(instruction, BR_SIMM19_START, SIMM9_LEN);
+            int64_t simm19 = getBitsAt(instruction, BR_SIMM19_START, SIMM9_LEN);
             int cond = getBitsAt(instruction, BR_COND_START, BR_COND_LEN);
             if (conditionCheck(cond, arm)) {
                 int64_t offset = simm19 * BYTES_IN_WORD;
